AG-13178 - Move legend into its own folder
