+++ packages/ag-charts-community/src/chart/factory/registerInbuiltModules.ts
@@ -6,7 +6,7 @@ import { LogAxis } from '../axis/logAxis';
 import { NumberAxis } from '../axis/numberAxis';
 import { TimeAxis } from '../axis/timeAxis';
 import { BackgroundModule } from '../background/backgroundModule';
-import { CommunityLegendModule } from '../legendModule';
+import { CommunityLegendModule } from '../legend/legendModule';
 import { NavigatorModule } from '../navigator/navigatorModule';
 import { AreaSeriesModule } from '../series/cartesian/areaSeriesModule';
 import { BarSeriesModule } from '../series/cartesian/barSeriesModule';
