+++ packages/ag-charts-community/src/chart/legend/legendDOMProxy.ts
@@ -1,19 +1,19 @@
-import type { LocaleManager } from '../locale/localeManager';
-import type { ModuleContext } from '../module/moduleContext';
-import type { Node } from '../scene/node';
-import type { Selection } from '../scene/selection';
-import { Transformable } from '../scene/transformable';
-import type { BBoxValues } from '../util/bboxinterface';
-import { DestroyFns } from '../util/destroy';
-import { createElement } from '../util/dom';
-import type { ButtonWidget } from '../widget/buttonWidget';
-import type { GroupWidget } from '../widget/groupWidget';
-import type { ListWidget } from '../widget/listWidget';
-import type { SwitchWidget } from '../widget/switchWidget';
-import type { Page } from './gridLayout';
+import type { LocaleManager } from '../../locale/localeManager';
+import type { ModuleContext } from '../../module/moduleContext';
+import type { Node } from '../../scene/node';
+import type { Selection } from '../../scene/selection';
+import { Transformable } from '../../scene/transformable';
+import type { BBoxValues } from '../../util/bboxinterface';
+import { DestroyFns } from '../../util/destroy';
+import { createElement } from '../../util/dom';
+import type { ButtonWidget } from '../../widget/buttonWidget';
+import type { GroupWidget } from '../../widget/groupWidget';
+import type { ListWidget } from '../../widget/listWidget';
+import type { SwitchWidget } from '../../widget/switchWidget';
+import type { Page } from '../gridLayout';
+import type { Pagination } from '../pagination/pagination';
 import type { CategoryLegendDatum } from './legendDatum';
 import type { LegendMarkerLabel } from './legendMarkerLabel';
-import type { Pagination } from './pagination/pagination';
 
 type ItemSelection = Selection<LegendMarkerLabel, CategoryLegendDatum>;
 type CategoryLegendDatumReader = { getItemLabel(datum: CategoryLegendDatum): string | undefined };
