+++ packages/ag-charts-community/src/chart/legend/legendDatum.ts
@@ -1,6 +1,6 @@
 import type { AgChartLegendListeners, AgMarkerShape } from 'ag-charts-types';
 
-import type { Scene } from '../scene/scene';
+import type { Scene } from '../../scene/scene';
 
 export interface ChartLegend {
     attachLegend(scene: Scene): void;
