+++ packages/ag-charts-community/src/module-support.ts
@@ -117,7 +117,7 @@ export { AxisTicks } from './chart/axis/axisTicks';
 export * from './chart/chartAxis';
 export * from './chart/crossline/crossLine';
 export * from './chart/crossline/crossLineLabelPosition';
-export * from './chart/legendDatum';
+export * from './chart/legend/legendDatum';
 export * from './chart/tooltip/tooltip';
 export * from './motion/animation';
 export * as Motion from './motion/easing';
@@ -175,7 +175,7 @@ export { OrdinalTimeScale } from './scale/ordinalTimeScale';
 export { Label } from './chart/label';
 export { Marker } from './chart/marker/marker';
 export { getMarker } from './chart/marker/util';
-export { LegendMarkerLabel } from './chart/legendMarkerLabel';
+export { LegendMarkerLabel } from './chart/legend/legendMarkerLabel';
 export { Diamond } from './chart/marker/diamond';
 export { Square } from './chart/marker/square';
 export { Triangle } from './chart/marker/triangle';
